requirements.txt
python-dotenv>=1.0.0

# LlamaIndex core
llama-index>=0.14.13
llama-index-core>=0.14.13
llama-index-workflows>=2.13.0

# Embeddings
llama-index-embeddings-huggingface>=0.6.1
sentence-transformers>=2.6.1
transformers>=4.41.0,<6.0.0
torch>=1.11.0
huggingface_hub>=0.23.0

# (Optional) OpenAI embeddings
llama-index-llms-openai>=0.6.0
llama-index-embeddings-openai>=0.5.0

# Vector store (Qdrant)
llama-index-vector-stores-qdrant>=0.9.0
qdrant-client>=1.7.0

# UI + PDF
streamlit>=1.32.0
pypdf>=4.0.0

# LLM (Gemini)
google-generativeai>=0.7.0

# Utilities
numpy
scikit-learn
scipy